+++ src/utils/babel/transform.test.ts
@@ -83,7 +83,7 @@ describe("transformHighlightContainingJsx", () => {
     });
   });
 
-  describe("When there is are multiple elements surrounded by strings", () => {
+  describe("When there are multiple elements surrounded by strings", () => {
     it("should return a transform result with the correct key, message and transformed code", () => {
       const input =
         'Bonjour de <a href="/foo">Vocab</a>!!! <Strong>Bonjour</Strong> Bonjour!';
@@ -93,6 +93,22 @@ describe("transformHighlightContainingJsx", () => {
         key: "Bonjour de Vocab!!! Bonjour Bonjour!",
         message: "Bonjour de <a>Vocab</a>!!! <Strong>Bonjour</Strong> Bonjour!",
         code: '{t("Bonjour de Vocab!!! Bonjour Bonjour!", { a: (children) => <a href="/foo">{children}</a>, Strong: (children) => <Strong>{children}</Strong> })}',
+      };
+
+      expect(result).toEqual(expected);
+    });
+  });
+
+  describe("When there are multiple of the same element", () => {
+    it("should suffix the duplicate keys in the hook call", () => {
+      const input =
+        'Bonjour de <a href="/foo">Vocab</a>!!! <a href="/bar">Bonjour</a> Bonjour!';
+
+      const result = transformHighlightContainingJsx(input);
+      const expected = {
+        key: "Bonjour de Vocab!!! Bonjour Bonjour!",
+        message: "Bonjour de <a>Vocab</a>!!! <a1>Bonjour</a1> Bonjour!",
+        code: '{t("Bonjour de Vocab!!! Bonjour Bonjour!", { a: (children) => <a href="/foo">{children}</a>, a1: (children) => <a href="/bar">{children}</a> })}',
       };
 
       expect(result).toEqual(expected);